Create unified RewardsTooltip component
