+++ apps/hyperdrive-trading/src/ui/markets/PoolRow/VariableApyStat.tsx
@@ -1,12 +1,4 @@
-import { fixed } from "@delvtech/fixed-point-wasm";
-import {
-  appConfig,
-  getHyperdriveConfig,
-  getToken,
-} from "@delvtech/hyperdrive-appconfig";
-import { SparklesIcon } from "@heroicons/react/16/solid";
-import { ChartBarIcon } from "@heroicons/react/24/solid";
-import * as Tooltip from "@radix-ui/react-tooltip";
+import { appConfig, getHyperdriveConfig } from "@delvtech/hyperdrive-appconfig";
 import { ReactNode } from "react";
 import Skeleton from "react-loading-skeleton";
 import { formatRate } from "src/base/formatRate";
@@ -15,6 +7,7 @@ import { GradientBadge } from "src/ui/base/components/GradientBadge";
 import { useIsNewPool } from "src/ui/hyperdrive/hooks/useIsNewPool";
 import { useCurrentLongPrice } from "src/ui/hyperdrive/longs/hooks/useCurrentLongPrice";
 import { PercentLabel } from "src/ui/markets/PoolRow/PercentLabel";
+import { RewardsTooltip } from "src/ui/rewards/RewardsTooltip";
 import { useRewards } from "src/ui/rewards/useRewards";
 import { useYieldSourceRate } from "src/ui/vaults/useYieldSourceRate";
 import { Address } from "viem";
@@ -63,168 +56,24 @@ export function VariableApyStat({
   }
 
   return (
-    <Tooltip.Provider>
-      <Tooltip.Root>
-        <Tooltip.Trigger className="flex items-center gap-2 whitespace-nowrap">
-          {yieldSourceRateStatus === "success" ? (
-            <PercentLabel
-              value={formatRate({
-                rate: yieldSourceRate?.netVaultRate ?? 0n,
-                includePercentSign: false,
-              })}
-              className="text-h4"
-            />
-          ) : (
-            <Skeleton width={100} />
-          )}
-          <GradientBadge>{multiplierLabel}</GradientBadge>⚡
-        </Tooltip.Trigger>
-        <Tooltip.Portal>
-          <Tooltip.Content
-            className="z-20 h-fit w-72 rounded-box bg-base-200 px-3 py-2 shadow-2xl"
-            sideOffset={5}
-            collisionPadding={12}
-          >
-            <div className="flex justify-between border-b border-neutral-content/30 p-3">
-              <p className="gradient-text text-lg">Rate & Rewards</p>
-            </div>
-            <div className="flex items-center justify-between border-b border-neutral-content/30 p-3">
-              <div className="flex items-center gap-1">
-                <ChartBarIcon className="h-4" />
-                Base APY
-              </div>
-              <div className="grid justify-items-end">
-                <p className="flex items-center gap-1">
-                  {(() => {
-                    if (isNewPool) {
-                      return (
-                        <>
-                          <SparklesIcon className="h-4" />
-                          New
-                        </>
-                      );
-                    }
-                    if (yieldSourceRate) {
-                      return fixed(yieldSourceRate.vaultRate).format({
-                        percent: true,
-                        decimals: 2,
-                      });
-                    }
-                    return "-";
-                  })()}
-                </p>
-              </div>
-            </div>
-
-            {appConfigRewards?.map((reward) => {
-              if (reward.type === "info") {
-                return (
-                  <div
-                    key={reward.iconUrl}
-                    className="flex flex-col items-start justify-start gap-2 border-b border-neutral-content/30 p-3 [&:nth-last-child(2)]:border-none"
-                  >
-                    <div className="flex items-center gap-4">
-                      <img
-                        src={reward.iconUrl}
-                        alt={`${reward.message}`}
-                        className="h-8"
-                      />
-                      <p>{reward.message}</p>
-                    </div>
-                  </div>
-                );
-              }
-              if (reward.type === "transferableToken") {
-                // safe to cast because we assume all rewards tokens are
-                // available in appConfig
-                const token = getToken({
-                  tokenAddress: reward.tokenAddress,
-                  chainId: reward.chainId,
-                  appConfig,
-                })!;
-
-                return (
-                  <div
-                    key={reward.tokenAddress}
-                    className="flex items-center justify-between border-b border-neutral-content/30 p-3 [&:nth-last-child(2)]:border-none"
-                  >
-                    <div className="flex items-center gap-1">
-                      <img
-                        src={token.iconUrl}
-                        alt={`${token.name} logo`}
-                        className="h-4"
-                      />
-                      {token.name}
-                    </div>
-
-                    <div className="grid justify-items-end">
-                      <p className="flex items-center gap-1">
-                        +
-                        {fixed(reward.apy).format({
-                          percent: true,
-                          decimals: 2,
-                        })}
-                      </p>
-                    </div>
-                  </div>
-                );
-              }
-
-              if (reward.type === "nonTransferableToken") {
-                // safe to cast because we assume all rewards tokens are
-                // available in appConfig
-                const token = getToken({
-                  tokenAddress: reward.tokenAddress,
-                  chainId: reward.chainId,
-                  appConfig,
-                })!;
-                return (
-                  <div
-                    key={reward.tokenAddress}
-                    className="flex items-center justify-between border-b border-neutral-content/30 p-3 [&:nth-last-child(2)]:border-none"
-                  >
-                    <div className="flex items-center gap-1">
-                      <img
-                        src={appConfig.protocols.morpho.iconUrl}
-                        className="h-4"
-                      />
-                      {token.name}
-                    </div>
-
-                    <div className="grid justify-items-end">
-                      <p className="flex items-center gap-1">
-                        +
-                        {fixed(reward.tokensPerThousandUsd).format({
-                          decimals: token.places,
-                        })}
-                      </p>
-                      <p className="text-2xs text-neutral-content">
-                        per $1000 / yr
-                      </p>
-                    </div>
-                  </div>
-                );
-              }
-            })}
-
-            <div className="flex items-center justify-between border-b border-neutral-content/30 p-3 [&:nth-last-child(2)]:border-none">
-              <div className="flex items-center gap-1">
-                <SparklesIcon className="h-4" />
-                Net APY
-              </div>
-
-              <div className="grid justify-items-end">
-                <p className="flex items-center gap-1">
-                  {fixed(yieldSourceRate?.netVaultRate || 0n).format({
-                    percent: true,
-                    decimals: 2,
-                  })}
-                </p>
-              </div>
-            </div>
-          </Tooltip.Content>
-        </Tooltip.Portal>
-      </Tooltip.Root>
-    </Tooltip.Provider>
+    <RewardsTooltip
+      chainId={chainId}
+      hyperdriveAddress={hyperdriveAddress}
+      baseRate={yieldSourceRate?.vaultRate}
+      netRate={yieldSourceRate?.netVaultRate}
+    >
+      {yieldSourceRateStatus === "success" ? (
+        <PercentLabel
+          value={formatRate({
+            rate: yieldSourceRate?.netVaultRate ?? 0n,
+            includePercentSign: false,
+          })}
+          className="text-h4"
+        />
+      ) : (
+        <Skeleton width={100} />
+      )}
+      <GradientBadge>{multiplierLabel}</GradientBadge>⚡
+    </RewardsTooltip>
   );
 }
